update log approved

+++ app/Http/Controllers/API/DeliveryController.php
@@ -46,8 +46,51 @@ class DeliveryController extends Controller
                     return $this->getResponse([],'Data gagal disimpan',500);
                 }
 
-                $deleteFile = DeliveryFiles::where('delivery_id',$delivery->id)->delete();
 
+                if(isset($request->file)){
+                    $deleteFile = DeliveryFiles::where('delivery_id',$delivery->id)->delete();
+                    foreach($request->file as $f){
+                        $data = (object) [
+                            'original' => $f,
+                            'delivery_id' => $delivery->id,
+                            'name' => $f->getClientOriginalName(),
+                            'extension' => $f->getClientOriginalExtension(),
+                            'is_image' => substr($f->getMimeType(), 0, 5) == 'image' ? true : false,
+                            'size' => $f->getSize(),
+                        ];
+                        $data->name = date('now').rand('10000','99999').'.'.$data->extension;
+
+                        $file = DeliveryFiles::create((array) $data);
+                        if(!$file){
+                            DeliveryFiles::where('delivery_id',$delivery->id)->forceDelete();
+                            DeliveryFiles::onlyTrashed()->where('delivery_id',$delivery->id)->restore();
+                            return $this->getResponse([],'File gagal disimpan',500);
+                        }
+
+                        $data->original->storeAs('delivery', $data->name);
+                    }
+
+                    DeliveryFiles::onlyTrashed()->where('delivery_id',$delivery->id)->delete();
+                }
+                return $this->getResponse(Delivery::with('delivery_files')->find($delivery->id),'Data berhasil disimpan',200);
+            }
+
+            $delivery = Delivery::firstWhere('transaction_id',$id);
+            if(!$delivery){
+                return $this->getResponse([],'Data tidak ditemukan',500);
+            }
+
+            $update = $delivery->update([
+                'estimation_date' => $request->estimation_date,
+                'description' => $request->description
+            ]);
+            if(!$update){
+                return $this->getResponse([],'Data gagal disimpan',500);
+            }
+
+
+            if(isset($request->file)){
+                $deleteFile = DeliveryFiles::where('delivery_id',$delivery->id)->delete();
                 foreach($request->file as $f){
                     $data = (object) [
                         'original' => $f,
@@ -69,47 +112,8 @@ class DeliveryController extends Controller
                     $data->original->storeAs('delivery', $data->name);
                 }
 
-                DeliveryFiles::onlyTrashed()->where('delivery_id',$delivery->id)->delete();
-                return $this->getResponse(Delivery::with('delivery_files')->find($delivery->id),'Data berhasil disimpan',200);
+                DeliveryFiles::onlyTrashed()->where('delivery_id',$delivery->id)->forceDelete();
             }
-
-            $delivery = Delivery::firstWhere('transaction_id',$id);
-            if(!$delivery){
-                return $this->getResponse([],'Data tidak ditemukan',500);
-            }
-
-            $update = $delivery->update([
-                'estimation_date' => $request->estimation_date,
-                'description' => $request->description
-            ]);
-            if(!$update){
-                return $this->getResponse([],'Data gagal disimpan',500);
-            }
-
-            $deleteFile = DeliveryFiles::where('delivery_id',$delivery->id)->delete();
-
-            foreach($request->file as $f){
-                $data = (object) [
-                    'original' => $f,
-                    'delivery_id' => $delivery->id,
-                    'name' => $f->getClientOriginalName(),
-                    'extension' => $f->getClientOriginalExtension(),
-                    'is_image' => substr($f->getMimeType(), 0, 5) == 'image' ? true : false,
-                    'size' => $f->getSize(),
-                ];
-                $data->name = date('now').rand('10000','99999').'.'.$data->extension;
-
-                $file = DeliveryFiles::create((array) $data);
-                if(!$file){
-                    DeliveryFiles::where('delivery_id',$delivery->id)->forceDelete();
-                    DeliveryFiles::onlyTrashed()->where('delivery_id',$delivery->id)->restore();
-                    return $this->getResponse([],'File gagal disimpan',500);
-                }
-
-                $data->original->storeAs('delivery', $data->name);
-            }
-
-            DeliveryFiles::onlyTrashed()->where('delivery_id',$delivery->id)->forceDelete();
             return $this->getResponse(Delivery::with('delivery_files')->find($delivery->id),'Data berhasil disimpan',200);
 
         }
